+++ sdwan_feature_profiles.tf
@@ -34,6 +34,12 @@ resource "sdwan_system_feature_profile" "system_feature_profile" {
   description = try(each.value.description, "")
 }
 
+resource "sdwan_topology_feature_profile" "topology_feature_profile" {
+  for_each    = { for t in try(local.feature_profiles.topology_profiles, {}) : t.name => t }
+  name        = each.value.name
+  description = try(each.value.description, "")
+}
+
 resource "sdwan_transport_feature_profile" "transport_feature_profile" {
   for_each    = { for t in try(local.feature_profiles.transport_profiles, {}) : t.name => t }
   name        = each.value.name